Move duplicate code in burst buffer plugins to a common location.
Reuse code to open the burst buffer state save file. Change the function
signature to accept a file name parameter so we can specify which state
save file to open, since the state save file to open will be different
for each plugin but the rest of the logic to open the state save file is
the same.