Allow threads to call the burst_buffer.lua script in parallel.
Calls to the burst_buffer.lua script need to be able to run in parallel
since they can last a long time. However, in order to run in parallel we
need a unique copy of the lua_State variable for every single time we
call the lua script. That means we need to load the lua script every
time we call it. Filesystem caching should make this a non-issue though.

Previously calls to burst_buffer.lua weren't protected by a mutex, so
there would have been problems with multiple threads calling this script
anyway.