srun - leave SLURM_DIST_UNKNOWN as default for --interactive
If the distribution wasn't specified we should not attempt to guess
block or cyclic for interactive step, since it's most likely only export
an environment variable that will be used by the next srun instead of
follow the logic for the step specific srun exec.

This was done for SallocDefaultCommand in eee6369503.

Bug 11400