NVML: Add AutoDetected unique_id for MIGs
unique_id is only set for MIGs, and is of the form:

    MIG-<GPU-UUID>/<GPU instance ID>/<compute instance ID>

unique_id is used to select from multiple MIGs via CUDA_VISIBLE_DEVICES
(i.e. GPU binding).

We do NOT use the MIG's UUID as the unique_id because we want to
support the lowest driver version. The MIG docs have this to say:

"1. With drivers >= R470 (470.42.01+), each MIG device is assigned a GPU
UUID starting with MIG-<UUID>.
"2. With drivers R450 and R460, each MIG device is enumerated by
specifying the CI and the corresponding parent GI. The format follows
this convention:
MIG-<GPU-UUID>/<GPU instance ID>/<compute instance ID>

"... While the rest of the examples in this document reference the old
enumeration format, the new format can be used equivalently."

https://docs.nvidia.com/datacenter/tesla/mig-user-guide/index.html#cuda-visible-devices

Bug 10970