GPUs: Use index instead of dev_num for CUDA_VISIBLE_DEVICES
Don't use dev_num (the number at the end of the device filename) since
this number (i.e. the minor number) is not guaranteed to be in order
of PCI bus ID, which is what CUDA_VISIBLE_DEVICES wants. Use index
instead, which _will_ be in PCI bus order if AutoDetect is used.
In the vast majority of cases, dev_num will match the PCI bus order,
but when it doesn't, things break.

Use device index, not dev_num, as the global GPU id in prolog/epilog.

NICs will still use dev_num as the global index to remain backwards
compatible.

NOTE: MPS also still uses dev_num, but should be converted over to use
index at some point.

Bug 10933