Testsuite - Test 39.18: Do not sort expected or actual output
The order of GPUs is important because it directly corresponds to the
bits of the GRES bitmaps used to track the GPUs, and this shouldn't
change unless we deliberately change it.

Also, we will eventually want to sort GPUs by PCI bus ID if AutoDetect
is used, so we need to test the order to make sure that works. We may
also want to guarantee that the GPU order in gres.conf is preserved if
AutoDetect is NOT used, so users can manually specify the PCI bus order.

Bug 8421