GRES: If MultipleFiles is specified, do not env bind to each file
Env vars like CUDA_VISIBLE_DEVICES specify indexes to physical GPUs
that were enumerated, NOT to individual device files. So stop setting
indexes to files and start setting indexes to the devices themselves,
or else you will get multiple entries for just a single physical
GPU. This is required to support MultipleFiles devices like NVIDIA MIGs.

Task binding (--gpu-bind) will alter usable_gres so that not all GPUs
will be bound to. When a GPU is skipped over due to usable_gres,
device_constrained needs to be set to true so that subsequent device
files are also skipped over and are not erroneously thought of as
separate devices.

Bug 10970