Wait UnkillableStepTimeout in proctrack_g_wait()
There was a hardcoded value of 256 seconds in proctrack_g_wait() in the
pgid and cgroup plugins that did not match the wait of the
step_terminate_monitor.

That caused values of UnkillableStepTimeout larger than 256 to not drain
the node when an unkillable step existed, because proctrack finished before
this 256 second timeout and the monitor was terminated.

With this fix, proctrack_g_wait() will wait UnkillableStepTimeout instead.

Also change the delay between tests to a maximum of 32 seconds, instead of
exponentially increasing it every time.

Bug 11103