Move task/cgroup memory task_cgroup_memory_create to new cgroup/v1
This is a big change that will remove all xcgroup calls from this function
into the new plugin. Since this one also starts the OOM monitor, I remove
the code which deals with it and just use the new code in cgroup/v1.

There's a new data struct for receiving information of OOM mem conditions
from the new plugin.