Fix issue with srun when running in a het allocation
The following example reproduces the problem:

salloc :
 srun : hostname

When initialize_and_process_args() is called, a check was made to
see if we had a het step and the existence of SLURM_JOB_NUM_NODES
was being used to determine if this was so with the assumption that
the variable was only ever set on a normal allocation (which was
not the case). The fix was to add two environment variable checks
to ensure that we have a het step: 1) SLURM_HET_SIZE must not exist
(otherwise we have an existing het allocation); and 2) SLURM_JOB_ID
must exist from an existing allocation (otherwise srun is creating
the allocation which is not a het step).

Bug 11271