Improve sacctmgr to avoid silently ignoring DefaultAccount
After last commit slurmdbd is not returning ESLURM_ACCESS_DENIED when
a Coordinator tries to implicitly set a DefaultAccount, but it's still
not letting Coordinators to set DefaultAccounts if user already has one.
From saccmgr we don't get any error message if we try to it explicitly,
this tries to avoid confusions.

Bug 9903