Fix coordinators when adding existing users with PrivateData=users
With PrivateData=users Coordinators are not able to see the users,
so they cannot see if a user that they want to add is already in
other accounts and with a DefaultAccount, but it should be able to
still add the user without changing the DefaultAccount.

Instead of returning ESLURM_ACCESS_DENIED, now slurmdbd overwrites
the is_def value if a Coordinator tries to set it for a user that
already has one.

Bug 9903