Avoid zombie threads.
Detach pthreads that are joinable but won't be joined.

Do this so that the thread resources will be freed when the thread
exits. If a joinable thread exits but is never joined, some resources
are still consumed by the system for each of these threads. See the
pthread_join man page.