AutoDetect: Sort GPUs by PCI bus order indirectly via links
When AutoDetect runs, do a final sort of GPUs by links after sorting by
device name.

Links will be set in order of GPU enumeration, which is in order of
PCI bus ID in the case of NVML. So at the end of AutoDetect, resort the
GPUs back in order of PCI bus ID based on links and the -1 value (which
indicates the spot of the current GPU).

Bug 10827