Do not use accrue limits for partition QOS
Because of multi-partition job submissions, accrue limits for partition
QOS didn't work properly. With this commit, we don't use partition QOS
when using accrue limits. We could make accrue limits work for multiple
partitions by ensuring the lowest limit on any partition QOS is used (or
use the job QOS if it has the flag OverPartQOS, or use assoc limits if
no QOS has a limit), which is what was previously happening but only for
a single partition QOS. However, we decided that might be confusing for
places that use multi-partition job sumissions since it could mean that
many jobs would not accrue age priority, and to our best knowledge right
now accrue limits are being used for job QOS.

Bug 7375