slurmd: move close for pidfile to the very end of execution
_kill_old_slurmd(), at the initialization stage, relies on this FD to know
when the old slurmd has finished. It was being closed before calling
fini() for the plugins, which then lead to a race between the new slurmd
init() and the shutdown slurmd's fini().

This was noticed with the new job_container/tmpfs due to how it handles
filesystem mount points.

Bug 11503