Fix creating system cgroups in stepd when not using core/memspec limit
Old cgroup code relied on the stepd checking for the existence of the
"system" cgroup, which should have been created by the slurmd, to
determine whether or not to add the pid to it or not.

New cgroup code didn't check for the "system" cgroup and just attempted
to create it in the stepd without being guarded by core/memspec
controls. Without a cgroup.conf, it would fail to create the "system"
cgroup.

Now the stepd won't attempt to create/use the "system" cgroup unless
core/memspec limit is configured.

Bug 9254