Don't reject a step if not enough nodes are available.
Consider this job:

    #!/bin/bash
    #SBATCH -N2
    srun -l -N1 bash -c 'hostname; sleep 10' &
    sleep 1
    srun -l -N2 bash -c 'hostname; sleep 10' &
    wait

The first step fills one node in the job allocation. The second step
is then submitted but is immediately rejected with a reason of
"Requested node configuration is not available".
This happens because the number of nodes available (1) is less than the
number of nodes requested (2), and we didn't set return_code in this
situation, so then return_code was set to
ESLURM_REQUESTED_NODE_CONFIG_UNAVAILABLE.

Before commit 6a2c99edbf96e, the second step wasn't rejected, the second
step incorrectly ran in parallel with the first step, thus overallocating
CPUs.

This patch makes the second step wait for enough nodes to be available to
run on. Once enough nodes are available for the second step, then it will
either run or it will either be rejected with the correct message if the
step request can't be fulfilled.

Bug 11857