Don't reject a step if it requests a node that is allocated.
Continuation of the previous commit.

Consider this job:

    #!/bin/bash
    #SBATCH -N2
    srun -l -N1 -wn1-1 bash -c 'hostname; sleep 10' &
    sleep 2
    srun -l -N2 -wn1-1 bash -c 'hostname; sleep 10' &
    wait

The second step requests a specific node (using srun -w) that is fully
allocated to the job. Without this commit, the second step is rejected
with the reason of "Memory required by task is not available".

Instead of rejecting the second step, this commit defers the second step
until the node it requested becomes available, and logs to slurmctld:

"some requested nodes n1-1 still have memory or CPUs used by other steps"

Bug 11857