Run PrologSlurmctld and EpilogSlurmctld in slurmscriptd.
By running PrologSlurmctld and EpilogSlurmctld in slurmscriptd, we avoid
fork() calls in slurmctld. Since slurmctld can potentially have a large
memory footprint and slurmscriptd is very lightweight, running these
scripts in slurmscriptd can be much faster simply because of the time it
takes for the fork() call to complete.

Bug 11896