Constrain cgroup devices on the task level
Do the required change to constrain devices when a task is started in a step
adding this task to the cgroup plugin and then constraining the gres plugin
provided devices. The job constrained devices are inherited since the task
plugin startup so we do not constrain it again per each task.

Note also that extern step pids are not added to a task, so we needed to
create a new function for this.

Bug 10375