Testsuite - Refactor 21.30.12 fixing existing race
Previously the test was not checking fully checking the aggregation of
multiple jobs to the GrpWall, just one.
Also it had a race with decay or precision problem leading to allow
the final job.
Now we submit N shorter jobs to reach the limit with 10% of extra time
and we check the Reason of the final job pending.

Bug 10656

Signed-off-by: Scott Jackson <scottmo@schedmd.com>