Forcing join order.
Using STRAIGHT_JOIN and putting the job_table at the end we make it so
we only need to look at so many records in the job_table which in
theory is larger than the assoc_table.  If the assoc_table is larger
this could result in slightly longer times.  But the assoc_table would
need to be > 100k to notice, a table that size is rather rare.

Bug 11742