Add SlurmScriptdPidFile to slurm.conf.
If slurmctld ever segfaulted, aborted, was killed by SIGKILL, or
experienced some other fatal error that caused it to skip the cleanup
code that happens on normal termination, slurmscriptd would be orphaned
and never killed. In order to prevent an orphaned slurmscriptd from
running forever after that point, create a SlurmScriptdPidFile that
saves the PID of slurmscriptd, and kill the old slurmscriptd whenever
slurmctld starts up immediately after we kill the old slurmctld.

Bug 11896