Indicate that we are destroying the step sooner.
There's a bunch of code that checks for g_step_cgpath being set, but when
we deleted the step we didn't clean it up. Do it now so further calls to
the cgroup plugin doesn't attempt to interact with it.

This caused errors when proctrack_g_get_pids() was received by the stepd
while cleaning up the freezer cgroup. Since the directory was not there
anymore the pids couldn't be obtained, and there was an error when
trying to open the directory.

Bug 11879