Do not add slurmstepd to the job cgroups
In the original code, the slurmstepd was not added to any cgroup except
freezer unless CoreSpecLimit, CoreSpecCount or MemSpecLimit were set.
During the refactor the proctrack logic was extended to all the controllers,
resulting in the stepd process being always added to the corresponding
containers.

Initially this seemed desirable. For example, PMI implements a key-value
database in the slurmstepd. A user process could thus make stepd consume
a considerable amount of memory, so constraining this was a good idea.

The issue is that adding the stepd to any cgroup means adding more logic
to add/remove the pid at the correct moment, and this was out of the scope
of this refactor. Also, it could cause undesired effects since the stepd
would now be constrained by job request, for example if a job just required
10MiB that could result in slurmstepd being killed by OOM.

This patch reverts this new logic, so the stepd forked by slurmd will reside
as before in the same cgroup than slurmd is, and will expect the admin to
set CoreSpecCount, CoreSpecLimit or MemSpecLimit in slurm.conf if he really
wants to control its memory or cpu resources.

We also avoid putting the stepd into the freezer container now. It seemed
not to buy anything, and all that provided is the need of logic to add and
remove the corresponding pid when the stepd cgroup was destroyed.

Bug 9254