Testsuite - alter approach to corrupting credentials.
Avoid the last 4 characters of the credential string. The ':' at the end,
as well as the trailing part of the base64-encoded string. It is possible
to alter the end of a base64 string and end up with an equivalent binary
stream after decoding, which can mean the attempt at corrupting the
credential will not succeed and it will decode properly, which caused
test 7.16 to fail randomly.

Bug 10276.