Docs - Fix AllowGroups partition parameter documentation
Documentation suggested that it's equivalent to PrivateData=jobs.
This is incorrrect since the data is just hidden from the SHOW_ALL
flag searches. Things like 'squeue -u <user>' will show jobs in the
partition, even if the caller is not permitted to run in that
partition because of AllowGroups.

Bug 11799