Add node_reg_mem_percent SlurmctldParameter
Allows a node to register with a percentage of memory without being
marked as invalid with low memory. The scheduler will schedule based on
the configured memory. If needed, a buffer of memory can be configured
with AllowedRamSpace. The task/cgroup plugin won't a memory limit
greater than the actual memory on the node.

Default is 95% for cloud nodes and 100% for everything else.

Bug 11533