burst_buffer/lua - Don't allow DefaultPool.
DefaultPool is currently documented that it only works for the
burst_buffer/datawarp plugin. Rather than try to make it work fully for
the burst_buffer/lua plugin, just don't allow DefaultPool. If
DefaultPool is set in burst_buffer.conf and burst_buffer/lua is in use,
an error message will be logged and DefaultPool will be unset.
Also do not use DefaultPool in burst_buffer/lua.