burst_buffer/lua - improve logging about the lua script.
Use the BurstBuffer debug flag to log the time it took to run a burst
buffer script rather than log at info if it took over an arbitrary amount
of time to run the script.

Reuse the burst buffer API name in a variable in logging where it is
used more than once.