Split step_partial_comp() in half to fix deadlock.
With 3651ba83443, kill_step_on_node() switch to list_for_each which will
hold the lock on the step List. step_partial_comp() needed that left
unlocked for the call to find_step_record().

But - re-locating the step_ptr when we have it already is pointless.
Split the portion of step_partial_comp() that looks it up away from
the part that doesn't require it.