burst_buffer - fix parsing the size and unit in bb_get_size_num().
Previously, if a pool and size (capacity) was given like this:

  capacity=1 pool=pool_name

then bb_get_size_num() would interpret the size as 1 and the unit as
" pool=pool_name" which is obviously wrong. This would result in a
failure in suffix_mult() which would print a debug message:

  Unrecognized numeric suffix ' pool=pool_name'

This commit fixes the use of strtoull() to get the size and optional
unit value.