Add spawned tasks to devices and cpuset cgroup
This is a continuation of commit 756207eaef. The original code of task
cgroup for cpuset and devices had one empty function each called
task_cgroup_[devices|cpuset]_attach_task, which were originally intended to
do a cgroup creation + adding the pid to the recently created cgroup.

These functions were empty because during the creation of the cgroup the
slurmstepd was implictly added to the corresponding cgroup, so this function
was effectively useless.

After commit 756207eaef we stopped implicitly adding the stepd to the
recently created cgroup, so the pid of the spawned tasks was not added and
remained unconstrained.

Note, this was not the case for task/cgroup memory or proctrack/cgroup
plugins where we were doing an explicit addition to the cgroup. This logic
showed again an inconsistency between different cgroup plugins.

This patch completes the work in 756207eaef, removing the implicit pid
addition in cpuset and devices, and adding an explicit call to the
corresponding add_pid, as it needs to be, making the plugins finally
consistent.

The call to add the spawned pid to the corresponding cgroup is done from
superior layers from _handle_add_extern_pid->task_g_add_pid, as one would
expect.

Bug 12039