Docs - MaxTime is applied to root/SlurmUser jobs
Update the doc to correct the statement - root and SlurmUser are not
allowed to submit a job with higher limits. If the job TimeLimit is
higher than configured MaxTime it's accepted or rejected at submission
based on EnforcePartLimits, regardless of the user. However, root,
SlurmUser and Operators are allowed to update the job TimeLimit above
the configured value.

Bug 12050