WIP: fix segfault when assigning pid to taskid
In mgr.c:_fork_all_tasks when assigning:

jobacct_id.taskid = job->task[i]->gtid + task_offset;

this jobacct_id.taskid doesn't match with the 'i' (i=0, taskid=1). Then we
call cgroup_g_task_addto which assumed that the taskid was addressable in
job->task[], but it is clearly not.

Get back to pass the pid, but investigation is needed to understand why
this taskid is not ok.

This obviously caused segfaults in cgroup_p_task_addto->_handle_task_cgroup.