Make slurmscriptd_run_bb_lua() wait for the response message.
When running this function we want it to block until the script is done
so we can get the return code and response message of the script.
However, slurmscriptd's current communication scheme is to use a single
set of anonymous pipes to communicate with slurmctld. This means we can't
block traffic on the anonymous pipe to wait for a response, because if we
did then no other communication could happen with slurmscriptd.

(If slurmscriptd opened a new socket for every connection then we could
obviously block on that socket, but it doesn't work that way right now.)

The solution was to use a hashmap, put waiting threads in the
hashmap, and pthread_cond_wait(). When slurmctld receives a
SLURMSCRIPTD_REQUEST_BB_LUA_COMPLETE RPC from slurmscriptd, slurmctld can
lookup the waiting thread in the hashmap and pthread_cond_signal() to
wake up that thread.