Call bb_g_run_script from slurmscriptd in a child process with timeout.
In slurmscriptd, fork() and call bb_g_run_script(). Poll the child
process to read a response message and get the exit status of the child.
run_command() was used as a model, except that instead of the child
exec()'ing a new program, the child calls bb_g_run_script(). If the
child times out, the parent (slurmscriptd) kills the child with
SIGTERM and then SIGKILL. track_script can also be used to kill the
child process.