Remove lua script tracking from track_script.
Continuation of last 2 commits. Because we are now using slurmscriptd to
run long-running burst_buffer.lua scripts, we aren't using track_script
anymore within the burst_buffer/lua plugin. This commit removes the code
that was added to be able to track a lua script running in a separate
thread (rather than a separate process). That code was added by the
commit:

    "Track threads running lua with track_script."

and this commit reverts much of that commit.