burst_buffer/lua - Set max values for time limits in burst_buffer.conf.
This plugin uses slurmscriptd to run some APIs in burst_buffer.lua so
that if they run longer than the configured timeout they can be killed.
SlurmScriptd looks for timeouts in milliseconds. In order to prevent
overflow when converting a 32-bit unsigned integer in seconds to a
32-bit signed integer in milliseconds, we set a maximum value for time
limits such that this conversion will never overflow.