burst_buffer/lua - Ignore ValidateTimeout.
burst_buffer.lua APIs that might use ValidateTimeout (slurm_bb_job_process
and slurm_bb_paths) are actually called directly from slurmctld, not
through SlurmScriptd. Because of this, they cannot be killed, so there
is no timeout for them. Therefore, ValidateTimeout doesn't matter in this
plugin, so this commit ignores whatever value is set for ValidateTimeout.