slurmscriptd - change process name and add log prefix.
slurmscriptd is fork()'d from slurmctld, so writes to the slurmctld log
file. To make debugging easier, set a prefix when logging in slurmscriptd
to distinguish it from slurmctld.

Also set the process name with setproctitle() so that when using "ps"
the slurmscriptd process is distinguishable from the slurmctld process.
And use prctl() to change the thread name so that LogTimeFormat=thread_id
will print "slurmscriptd".

Bug 12093