Revert "Add SlurmScriptdPidFile to slurm.conf."
This reverts commit 4cf430dd483953bf93a7f6b64c0af6d116b353b4.

SlurmScriptdPidFile was originally added to make sure that slurmscriptd
was cleaned up even if slurmctld died in an unexpected way. However, we
can't make SlurmScriptdPidFile in the default location because the
process is SlurmUser, which doesn't have permissions to write in the
default location for that file. There isn't a clean way to fix this, so
we will simply remove SlurmScriptdPidFile and ensure that slurmscriptd
gets cleaned up in a different way.

Bug 12097