slurmd - fix assert failure on initialization due to bad node name.
Move the already existing find_node_record2() call up in the code and
bail with error if no record found.

The justification for the call movement is to avoid unneeded spool dir
creation attempt as well as further operations for non-configless setups
if no record is found.

Change the used call to find_node_record() too while here so that the
function itself logs the error if no record found for us.

Bug 11621