burst_buffer/lua - don't throttle highest priority jobs in stage_in.
Continuation of the previous commit.

For stage_in we need to run burst_buffer.lua in job priority order so
that highest priority jobs can start as soon as possible. The previous
commit throttled burst buffer staging but didn't guarantee the order in
which the throttled threads would run. This commit only queues as many
stage in threads as we are allowed so we don't need to throttle stage in
threads.