Don't close a file descriptor that wasn't allocated.
bb_job->memfd was initialized to zero when bb_job was xmalloc'd. We only
set bb_job->memfd if burst buffer symbol replacement is used in the job
script. However, in _bb_job_del2 we always called close() on bb_job->memfd
which may have never been set, and if it wasn't set then we close file
descriptor zero which is a valid file. That could cause strange issues in
other unrelated places dealing with files.

We never hit this error because we never actually called bb_job_del()
after it is successfully allocated. When we found and tried to fix this
memory leak by calling bb_job_del(), we found this issue.