Continuation of bf2e21f9d40d8.
Turns out scontrol reconfig calls this as well which messes things up.
It also happens that --exclusive doesn't work correctly with bf2e21f.

One thing we know for sure if the alloc_list isn't created yet this is
a new allocation so we will base things off that.

Bug 12123