Check that the step requests at least as many gres as nodes.
This just checks that the step requested at least as many gres as
nodes. Without this check, it is possible for a step to be allocated one
GPU and two nodes, which results in one node having zero GPUs. We don't
want to allow this with normal steps because (1) we want to be
consistent with jobs, and this isn't allowed for job allocations, and
(2) this would make it impossible to calculate --mem-per-gpu on the node
with zero GPUs.

This isn't a perfect check since it just checks the request, not what
is eventually allocated. Although srun always seems to set min_nodes,
it is possible to manually set min_nodes to zero in the RPC layer or to
manually modify srun to set min_nodes to zero. In that case this check
fails.

Bug 12203