Log error to user when step allocation fails due to gres counts.
If a step requests GRES and it fails some basic tests (step requests
more GRES than the job requested) then print an error message in the
shell where srun was invoked. This helps the user know exactly why the
GRES specification was rejected.

This commit only sets this error message in one spot, but it could be
used in other places during step allocation.

Bug 12203