Add SLURM_JOB_GPUS back into Prolog; add it to Epilog
SLURM_JOB_GPUS was effectively removed from the Prolog in 19.05 with
commit deaf17ab0db3 (the prolog no longer used the same function to set
the env as the job, and the new function forgot to add in
SLURM_JOB_GPUS). So re-add it to 21.08.
Adding it to the Prolog in gres_p_epilog_set_env() also adds it to the
Epilog, so update the docs to reflect this.

Bug 8596