Pass additional parameters to job_get_tres_mem().
They aren't used in this commit but will be used to estimate memory based
on mem_per_gres and number of gres requested.

Also, since we modified the comment above the function in the header
file, remove the comment above the function in the C source file to
avoid these comments getting out of sync.

Bug 11837