Changed ReqMem field in sacct to match memory from ReqTRES.
ReqMem now shows the requested memory of the whole job with a letter
appended indicating units (M for megabytes, G for gigabytes, etc.). ReqMem
is only displayed for the job, since the step does not have requested
TRES. Previously ReqMem was also displayed for the step but was just
displaying ReqMem for the job, so it was wrong.

Bug 11837