Update configs_for_slurmd.config_files when issuing reconfig.
When running in configless, scontrol reconfig would result in empty
configs being sent to the slurmd and a segfault.

This now needs to support three potential slurmd versions, so
explicitly handle each version. Long-term, this will hopefully
collapse back down to a single unified protocol, but at the moment
this is the only way to support a mix of slurmd versions correctly.

Bug 12188