Remove overzealous GRES type check in the controller
Remove check on whether all GRES for each GRES name is either all typed
or all not across all nodes when registering with the controller.

There are cases where one node could have a typed GPU and another
node could have an untyped GPU. This is a valid configuration, and
should not emit an error. This can happen if explicitly configured, or
when an AutoDetect/conf or slurm.conf/gres.conf merger discards the TYPE
flag when discarding the type.

Bug 11097