Fix slurmstepd SPANK prolog/epilog handler use-before-init issue.
The read_slurmd_conf_lite() function reads slurm_conf values but the struct
has not yet been initialized. This was detected because the LogTimeFormat
was different when logs were logged from the job_script context.

Moving the call to slurm_conf_init() before read_slurmd_conf_lite()
solves the issue. After commit 90093d0878, log_fmt is not
packed/unpacked anymore as part of the slurmd_conf_t (slurmd conf lite).
Same happens with job_acct_gather_freq after commit 236be1f374, but this
field is not that relevant for the stepd when serving as job_script context.

For future reference, slurm_conf_init() is already called in slurmstepd
main() but it is called after _process_cmdline(). It could potentially be
moved to before _process_cmdline() and remove the other call in
_handle_spank_mode(), but perhaps it is done this way to avoid initializing
the config when the stepd is used for "getenv" (env_array_user_default()).

Bug 8779