Start burst buffer stage-in for as many jobs in a job array as possible.
Previously we only started stage in for one job in a job array per
scheduling cycle, even for the jobs that were split from the job array.
This commit fixes logic in both the main and backfill schedulers so that
all the jobs that have been split from the job array may attempt to start
burst buffer stage-in per scheduling cycle.

Bug 12202