Clarify and add comments in GRES dealloc code
We actually *do* want to clear job GRES bits if we aren't doing
scheduling tests and we are resizing.

I think "emulated deallocate" was referring to job testing, where node
bitmaps are altered temporarily and then reverted after the job test. In
that case, we don't want to mess with the job's GRES.

Bug 11881