Implement constrain_get and set
These functions are quite similar to the v1 ones but there are important
changes in the interfaces we interact with. Also there are not swappiness
anymore, devices, or kmemspace either.

Also, there we are not implementing some unused constrain_get or _set. For
example a constrain_get for memory controller is not needed at the moment.

Bug 12255