Require requesting a GPU if --mem-per-gpu is requested.
Previously it was possible to request --mem-per-gpu but not request any
GPUs. The result of this was the job was not allocated any memory but
would still run. On the compute node, the job would be given the node's
entire physical memory in the memory cgroup.

To prevent this, we require that at least one GPU must be requested if
--mem-per-tres is requested.

Bug 12284