Properly resize a job's GRES bitmaps and counts when resizing the job.
Resize the job's GRES bitmaps and counts for select/cons_[t]res and
select/linear (before, only the node's GRES bitmaps were being altered
on resize).

Add `resize` in gres_ctld_job_dealloc() to indicate when a GRES
dealloc is due to a resize.

If a job is resizing instead of finishing, then the job's GRES will
stick around. So we need to actually update the job's GRES and shift
all of the bitmaps to fill the hole caused by removing all the GRES on
the node being removed.

NEWS is for last 3 commits as well.

Bug 11881